requirements.txt
biopython==1.63
binaryornot==0.3.0
PyYAML==3.11
click==3.3